Refactor: MaxBackups removed from game settings. Refactor: Savegame backup system is now self-contained like the savegame slot system. (user has to keep track of values like current backup number, etc. themselves) Refactor: Savegame slot and savegame backup system moved to partial class GameSavegameSystem.cs
